requirements.txt
scrapy
cryptography==0.5.4
ipython
scipy
scikit-learn
pandas
git+git://github.com/vishnubob/python-midi.git@HEAD#egg=midi
scikits.audiolab